requirements.txt
azure-identity
azure-mgmt-costmanagement
pandas
azure-mgmt-resource